pricebackers_shopify_app/requirements.txt
asgiref==3.2.10
Django==3.0.8
pyactiveresource==2.2.2
pytz==2020.1
PyYAML==5.3.1
ShopifyAPI==8.0.1
six==1.15.0
sqlparse==0.3.1
python-dotenv
requests
djangorestframework
dj_rest_auth==2.2.4
django-allauth==0.50.0
django-cors-headers
python-decouple==3.6
psycopg2
uWSGI>=2.0.19.1,<2.1
